requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
pyyaml